chore: Clean-up unused MT lookup strategies
